adding command autocomplete, per channel config and module changed default HTTPQuery from POST to GET sorted bot map fixed factoid forgeting and added unforget, moved factoid class to main jar
probably more stuff. should problem review changes. some things likely will be removed in near future

+++ src/pl/shockah/shocky/sql/CriterionNumber.java
@@ -1,7 +1,7 @@
 package pl.shockah.shocky.sql;
 
 public class CriterionNumber extends Criterion {
-	public CriterionNumber(String column, Operation o, int value) {
+	public CriterionNumber(String column, Operation o, long value) {
 		super(column+o+value);
 	}
